minor fix to get rid of warnings
git-svn-id: http://svn.civicrm.org/civicrm/trunk@3761 cff41f8d-9fe8-0310-9d4e-fb156a3ffe5e

+++ xml/GenCode.php
@@ -42,9 +42,11 @@ $smarty =& new Smarty( );
 $smarty->template_dir = './templates';
 
 $compileDir = '/tmp/templates_c';
-$oldTemplates = preg_grep('/tpl\.php$/', scandir($compileDir));
-foreach ($oldTemplates as $templateFile) {
-    unlink($compileDir . '/' . $templateFile);
+if (file_exists($compileDir)) {
+    $oldTemplates = preg_grep('/tpl\.php$/', scandir($compileDir));
+    foreach ($oldTemplates as $templateFile) {
+        unlink($compileDir . '/' . $templateFile);
+    }
 }
 $smarty->compile_dir = $compileDir;
